+++ src/components/BatchAllocations/CreateBatchAllocations/CreateBatchAllocations.test.js
@@ -13,6 +13,7 @@ import { useFiscalYear } from '../../../common/hooks';
 import {
   useSourceData,
   useBatchAllocation,
+  useBatchAllocationFormHandler,
   useBatchAllocationMutation,
 } from '../hooks';
 import { CreateBatchAllocations } from './CreateBatchAllocations';
@@ -29,6 +30,7 @@ jest.mock('../hooks', () => ({
   ...jest.requireActual('../hooks'),
   useSourceData: jest.fn(),
   useBatchAllocation: jest.fn(),
+  useBatchAllocationFormHandler: jest.fn(),
   useBatchAllocationMutation: jest.fn(),
 }));
 
@@ -53,15 +55,17 @@ const renderComponent = (props = {}) => render(
 );
 
 describe('CreateBatchAllocations', () => {
-  const recalculate = jest.fn(() => Promise.resolve({}));
+  const recalculate = jest.fn(() => Promise.resolve({ fyFinanceData }));
   const batchAllocate = jest.fn();
   const showCallout = jest.fn();
+  const handle = jest.fn(() => Promise.resolve());
 
   beforeEach(() => {
     useBatchAllocation.mockReturnValue({ budgetsFunds: fyFinanceData, isLoading: false, refetch: () => {} });
     useSourceData.mockReturnValue({ data: { name: 'Source Data' } });
     useFiscalYear.mockReturnValue({ fiscalYear: { code: '2025' } });
     useBatchAllocationMutation.mockReturnValue({ recalculate, batchAllocate });
+    useBatchAllocationFormHandler.mockReturnValue({ handle });
     useShowCallout.mockReturnValue(showCallout);
   });
 
@@ -85,6 +89,20 @@ describe('CreateBatchAllocations', () => {
     await userEvent.click(screen.getByRole('button', { name: 'ui-finance.allocation.batch.form.footer.recalculate' }));
 
     expect(recalculate).toHaveBeenCalled();
+  });
+
+  it('should call handle on submit', async () => {
+    renderComponent();
+
+    await waitFor(() => {
+      expect(screen.getByText('ui-finance.allocation.batch.form.title.edit')).toBeInTheDocument();
+    });
+
+    await userEvent.type(screen.getAllByLabelText('ui-finance.transaction.allocation.batch.columns.budgetAllocationChange')[0], '42');
+    await userEvent.click(screen.getByRole('button', { name: 'ui-finance.allocation.batch.form.footer.recalculate' }));
+    await userEvent.click(screen.getByRole('button', { name: 'stripes-components.saveAndClose' }));
+
+    expect(handle).toHaveBeenCalled();
   });
 
   it('should handle recalculate errors', async () => {