+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Migrate to shared GA workflows. Refs UIF-572.
 * Fixes for batch allocation form. Refs UIF-579.
 * Hide future FYs from group details view. Refs UIF-587.
+* Save and close batch allocations. Refs UIF-534.
 * Search and filter batch allocation logs. Refs UIF-541.
 
 ## [7.0.1](https://github.com/folio-org/ui-finance/tree/v7.0.1) (2024-12-02)